split RollingUpdateST into two suites
